Javadoc cleanup
