+++ Source/Core/src/ca/uqac/lif/dag/LabelledNode.java
@@ -1,5 +1,26 @@
+/*
+    Petit Poucet, a library for tracking links between objects.
+    Copyright (C) 2016-2021 Sylvain Hallé
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Lesser General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU Lesser General Public License
+    along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
 package ca.uqac.lif.dag;
 
+/**
+ * A node that contains an object, called its "label".
+ * @author Sylvain Hallé
+ */
 public class LabelledNode extends Node
 {
 	/**
@@ -7,6 +28,10 @@ public class LabelledNode extends Node
 	 */
 	/*@ null @*/ protected Object m_label;
 	
+	/**
+	 * Creates a new labelled node.
+	 * @param label The node's label
+	 */
 	public LabelledNode(/*@ null @*/ Object label)
 	{
 		super(1, 1);
@@ -22,7 +47,6 @@ public class LabelledNode extends Node
 		}
 		return "null";
 	}
-	
 	
 	/**
 	 * Connects the first output pin of the current node to the first input pin
